Modifiche per funzionare con Windows fatte, rimangono tanti todo di miglioramento ma la base funziona e aiuta

+++ tritalaravel.php
@@ -3,11 +3,18 @@
 include "lib/functions.php";
 
 //config
+$with_new_db = true; // creare db nuovo per ogni utente o sempre lo stesso?
+$with_npm_build = true; // lanciare anche npm install e build?
+$windows = true;
+
 $dbName = "tritalaravel"; //quest
 $config_dir = __DIR__ . "/config";
-$repos_dir = "tritalaravel"; //se cambiata da mettere in .gitignore
-$repo_name = "laravel-many-to-many";
+$folder_name = "tritalaravel"; //se cambiata da mettere in .gitignore
+$repo_name = "laravel-comics";
 //fine config
+
+// @todo switch case con scelta browser come scritta ma selezione poi del percorso
+$browser = 'C:\Program Files\Mozilla Firefox\firefox.exe';
 
 echo "*** Inizio verifica compiti ***\n";
 
@@ -54,14 +61,13 @@ $processFirstHalf = 2; // Imposta su false per processare la seconda metà
 
 // Determina la lunghezza dell'array e calcola il punto di divisione
 $userCount = count($users);
-$halfPoint = ceil($userCount / 2);
-
+$halfPoint = 17; // Francesco Mascellino esattamente ... anche se Davide e Lorenzo han abbandonato // ceil($userCount / 2);
 
 echo "Studenti $userCount\n";
 
-if (!file_exists($repos_dir)) {
+if ( ! file_exists($folder_name)) {
     echo "Creazione directory delle reopository necessaria\n";
-    mkdir($repos_dir, 0777, true);
+    mkdir($folder_name, 0777, true);
 }
 
 echo "Directory repository OK\n";
@@ -72,11 +78,12 @@ while (true) {
 
     echo "Elenco degli utenti:\n";
     foreach ($users as $index => $user) {
-        echo ($index + 1) . ". $user\n";
+        echo ($index) . ". $user\n";
     }
 
     // Chiedi quale/i utente/i verificare
-    $response = readline("Quale utente vuoi verificare? (Inserisci un numero, 'da [numero]', 'prima metà', 'seconda metà', 'exit' per uscire): ");
+    echo "Quale utente vuoi verificare? (Inserisci un numero, 'da [numero]', 'prima metà', 'seconda metà', 'exit' per uscire): ";
+    $response = fgets(STDIN);
 
     // Gestisci il comando di uscita
     if (strtolower($response) === 'exit') {
@@ -87,7 +94,7 @@ while (true) {
     // Gestisci la risposta
     if (is_numeric($response) && $response > 0 && $response <= $userCount) {
         // Un solo utente specificato
-        $selectedUsers = array_slice($users, $response - 1, 1);
+        $selectedUsers = array_slice($users, $response, 1);
     } elseif (strpos($response, 'da ') === 0) {
         // Intervallo specificato
         $start = intval(substr($response, 3)) - 1;
@@ -101,12 +108,16 @@ while (true) {
     }
 
     foreach ($selectedUsers as $user) {
-
-        $user_dir = "$initial_dir/$repos_dir/$user";
+        $user_dir = "$initial_dir/$folder_name/$user";
         $repo_dir = "$user_dir/$repo_name";
 
-        echo "Eliminazione vecchio db $dbName\n";
-        executeQuery("DROP DATABASE `$dbName`;");
+        // @todo new Database?
+        if ($with_new_db) {
+            $dbName .= '_' . $user;
+        } else {
+            echo "Eliminazione vecchio db $dbName\n";
+            executeQuery("DROP DATABASE IF EXISTS `$dbName`;");
+        }
 
         echo "Controllo sporcizia $user\n";
         if (file_exists($user_dir)) {
@@ -117,7 +128,7 @@ while (true) {
         }
 
         echo "Creazione nuovo db $dbName\n";
-        executeQuery("CREATE DATABASE `$dbName`;");
+        executeQuery("CREATE DATABASE IF NOT EXISTS `$dbName`;");
 
         echo "Creazione nuova directory $dbName $user_dir\n";
         mkdir($user_dir, 0777, true);
@@ -140,6 +151,9 @@ while (true) {
         $envContent = file_get_contents($config_dir . "/.env", true);
         echo "Salvataggio .env per $user\n";
         // Scrivi il contenuto modificato nel file .env
+        if ($with_new_db) {
+            $envContent = str_replace('DB_DATABASE=tritalaravel', 'DB_DATABASE=' . $dbName, $envContent);
+        }
         file_put_contents($repo_dir . "/.env", $envContent);
 
         // Installazione di npm
@@ -147,58 +161,75 @@ while (true) {
         run("npm install");
 
         echo "/usr/local/bin/php /Users/mistre/Library/Application\ Support/Herd/bin/composer install\n";
-        run("/usr/local/bin/php /Users/mistre/Library/Application\ Support/Herd/bin/composer install"); // Installazione delle dipendenze Composer
+        run("composer install"); // Installazione delle dipendenze Composer
+
         echo "/usr/local/bin/php artisan key:generate \n";
-        run("/usr/local/bin/php artisan key:generate");
+        run("php artisan key:generate");
+
         echo "/usr/local/bin/php artisan migrate:install \n";
-        run("/usr/local/bin/php artisan migrate:install");
+        run("php artisan migrate:install");
+
         echo "/usr/local/bin/php artisan migrate \n";
-        run("/usr/local/bin/php artisan migrate");
+        run("php artisan migrate");
+
         echo "/usr/local/bin/php artisan db:seed \n";
-        run("/usr/local/bin/php artisan db:seed");
+        run("php artisan db:seed");
+
         echo "/usr/local/bin/php artisan storage:link \n";
-        run("/usr/local/bin/php artisan storage:link");
+        run("php artisan storage:link");
 
         echo "Creazione utente di test user:tritalaravel@boolean.it pass:tritalaravel\n";
         $email = 'tritalaravel@boolean.it';
-        $password = password_hash('tritalaravel', PASSWORD_DEFAULT); // Hash della password
-        $emailVerifiedAt = date('Y-m-d H:i:s'); // Timestamp corrente per email verificata
+        $password = password_hash('tritalaravel', PASSWORD_DEFAULT); // Hash della password: $2y$10$AGMGqtB0TjrZCjxjlQWVDuYVGV8dpGTU5Q7zeiGAlhCNqDRHXkH9i
+        // replaced with , CURRENT_TIMESTAMP $emailVerifiedAt = date('Y-m-d H:i:s'); // Timestamp corrente per email verificata
 
         $query = "INSERT INTO 
               users (name, email, password, email_verified_at) 
-              VALUES ('Paolo', '$email', '$password', '$emailVerifiedAt');
+              VALUES ('Paolo Mistretta', '$email', '$password', CURRENT_TIMESTAMP);
              ";
 
         executeQuery($query, $dbName);
 
         echo "npm run dev & $\n";
-        $npmProcess = run("npm run dev > /dev/null 2>&1 & echo $!"); // Avvio NPM in background
+        //$npmProcess = run("npm run dev > /dev/null 2>&1 & echo $!"); // Avvio NPM in background
+        if ($with_npm_build) {
+            $npmProcess = run("npm run build"); // Build di NPM
+        } else {
+            $npmProcess = run("npm run dev"); // Avvio NPM in background
+        }
         echo "/usr/local/bin/php artisan serve &\n";
-        $laravelProcess = run("/usr/local/bin/php artisan serve > /dev/null 2>&1 & echo $!"); // Avvio Laravel in background
+        //$laravelProcess = run("/usr/local/bin/php artisan serve > /dev/null 2>&1 & echo $!"); // Avvio Laravel in background
+        $laravelProcess = run("php artisan serve"); // Avvio Laravel in background
 
         // Avvio di Chrome
-        echo "open -a 'Google Chrome' http://localhost:8000 \n";
-        run("open -a 'Google Chrome' http://localhost:8000"); // Sostituire con il comando appropriato per il tuo sistema operativo
+        echo "open/start Browser at http://localhost:8000 \n";
+        if ($windows) {
+            run("start '{$browser}' 'http://127.0.0.1:8000'");
+        } else {
+            run("open -a 'Google Chrome' http://localhost:8000"); // Sostituire con il comando appropriato per il tuo sistema operativo
+        }
 
         // Attesa della chiusura di Chrome
         echo "Valutazione umana di $user\n";
-        readline("Premi [Enter] dopo aver chiuso Chrome...");
+        echo "Premi [d] per eliminare tutto, [Qualsiasi tasto] per uscire e basta";
+        $press = fgets(STDIN);
 
-        // Terminazione dei processi Laravel e NPM
-        echo "Terminazione dei processi per $user\n";
-        run("kill $laravelProcess");
-        run("kill $npmProcess");
+        if ($press = 'd') {
+            // Terminazione dei processi Laravel e NPM
+            echo "Terminazione dei processi per $user\n";
+            run("kill $laravelProcess");
+            run("kill $npmProcess");
 
-        echo "Eliminazione db $dbName\n";
-        executeQuery("DROP DATABASE `$dbName`;");
+            echo "Eliminazione db $dbName\n";
+            executeQuery("DROP DATABASE `$dbName`;");
 
-        echo "Cancellazione directory $repo_dir\n";
-        run("rm -rf $repo_dir");
+            echo "Cancellazione directory $repo_dir\n";
+            run("rm -rf $repo_dir");
 
-        // Torna alla directory corrente iniziale dopo ogni operazione
-        chdir($initial_dir);
+            // Torna alla directory corrente iniziale dopo ogni operazione
+            chdir($initial_dir);
+        }
     }
 }
-
 
 ?>
